Add s390x build support + documentation (#5)
* Build s390x gomplate in a Docker container

Gomplate is required for running the Instana agent in a Docker
container. There are no pre-built gomplate binaries for
Ubuntu 18.04 s390x. So build it from source with the script
"build_s390x_gomplate.sh" which creates a Docker image
"gomplate:latest", runs it as "gomplate", copies the gomplate
binary from the container to this project directory, and cleans up
container and image again.

* Add s390x Instana agent image build

Add the file Dockerfile.s390x. It differs from the x86_64 Dockerfile
in the following points:
* architecture is set to s390x
* docker-ce package is used instead of the docker-ce-cli
* gomplate binary is added from this project directory

The newer docker-ce-cli package which replaces docker-ce is plain
not provided for s390x on the docker website.
For the gomplate binary run ./build_s390x_gomplate.sh first.

Build like this:
docker build ./ --build-arg FTP_PROXY=${INSTANA_AGENT_KEY} \
  -f Dockerfile.s390x -t instana_agent_s390x:latest --no-cache

Run like this:
docker run --name instana_agent -h instana_agent_host -dt \
  -e INSTANA_AGENT_KEY="..." -e INSTANA_AGENT_ENDPOINT="..." \
  -e INSTANA_AGENT_ENDPOINT_PORT="..." \
  -v "/opt/ibm/java-s390x-80/:/opt/instana/agent/jvm" \
  instana_agent_s390x:latest

On the host system it is required that you have installed the IBM
Java SDK for Linux on z Systems 64-bit to /opt/ibm/java-s390x-80/.
The source link is https://developer.ibm.com/javasdk/downloads/sdk8/
and an example installer file name is ibm-java-s390x-sdk-8.0-5.35.bin.

* README: Add s390x build documentation

The s390x build is special. The customers should do it themselves.
So more detailed documentation is required. Add it as
README_s390x.md, reference it in README.md, and also fix some
formatting in README.md for better readability.

* s390x: Use and verify gomplate release 3.4.1

As a review result of #5, fetch the latest numbered release of
gomplate (v3.4.1) instead of using the master build. Also verify
a correct download of the source archive with sha256sum.
Use tar instead of unzip to get rid of a further dependency and
use the long option --version of gomplate for better readability.